New features        : Problems solved     : - FIXED: I32/I64 declaration problem.
git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@7690 4b12f95d-9313-0410-afc9-bd26427041f1

+++ area51/cc2doptcontrol2/src/spacematvecassembly.f90
@@ -908,7 +908,7 @@ contains
   
     ! local variables
     real(dp), dimension(:), pointer :: p_Ddata
-    integer(i32), dimension(:), allocatable :: p_Idofs
+    integer, dimension(:), allocatable :: p_Idofs
     integer :: i,nviolate
     real(dp) :: du
